Fixed cpu rendering scale being incorrect.
From Rinnegatamante

+++ Source/SysCTR/Graphics/NativeTextureCTR.cpp
@@ -67,8 +67,8 @@ CNativeTexture::CNativeTexture( u32 w, u32 h, ETextureFormat texture_format )
 ,	mpPalette( NULL )
 ,	mTextureId( 0 )
 {
-	mScale.x = 1.0f / mCorrectedWidth;
-	mScale.y = 1.0f / mCorrectedHeight;
+	mScale.x = 1.0f / (float)mCorrectedWidth;
+	mScale.y = 1.0f / (float)mCorrectedHeight;
 	
 	glGenTextures( 1, &mTextureId );
